+++ CHANGES.rst
@@ -8,13 +8,17 @@ New Tools and Services
 API changes
 -----------
 
-
 esa.euclid
 ^^^^^^^^^^^^
 
 - New module to access the ESA Euclid Archive. [#3216]
 
 - Tests remove temp directories. [#3226]
+
+ipac.irsa
+^^^^^^^^^
+
+- ``query_sia`` now returns the results as an astropy QTable. [#3252]
 
 mast
 ^^^^